+++ business/src/index.ts
@@ -12,4 +12,5 @@ export * from "./operations";
 export * from "./adapters";
 export * from "./chart-adapters";
 export * from "./editor-adapters";
+export * from "./data-source";
 export { registerBusinessElements } from "./register";